Refactored the physics folder structure
