+++ src/jaff/physics/__init__.py
@@ -1,7 +1,11 @@
 from . import constants
 from ._equations import get_sfluxes, get_sodes, get_sradodes
-from ._photochemistry import Photochemistry
-from ._radiation import Radiation, RadiationGroup, RadiationGroupReactionProps
+from .photo_reactions._photochemistry import Photochemistry
+from .photo_reactions._radiation import (
+    Radiation,
+    RadiationGroup,
+    RadiationGroupReactionProps,
+)
 
 __all__ = [
     constants,
