+++ src/jaff/physics/_typing/__init__.py
@@ -1,4 +1,3 @@
 from ._equations import RadiationGroupReactionProps
-from ._photochemistry import XsecsProps
 
-__all__ = [RadiationGroupReactionProps, XsecsProps]
+__all__ = [RadiationGroupReactionProps]
