+++ src/jaff/physics/photo_reactions/_typing/__init__.py
@@ -1,0 +1,3 @@
+from ._photochemistry import XsecsProps
+
+__all__ = [XsecsProps]
