+++ src/jaff/physics/photo_reactions/_typing/_photochemistry.py
